Updated events widget and page to display events correctly.
Specifically, events are now shown in reverse chronological order. This means in the upcoming events widget events are shown in current-then-soonest-event-first order. And on the events page, past events are shown in most-recent-first order. Also, these lists of events are populated using JavaScript starting from the entire set of events present when the site was built. This means events will automatically be moved from upcoming to past when they are done without having to rebuild the site.