Fixed bug in default layout
The main div in the default layout was using flex so it would play nice with the card column. This was breaking the width of the nested container div so that element was using flex to maximimize its width. This this was causing the nested container to expand beyond the width of the main div, which was not desired. To fix the issue, the card column was moved inside the nested container div and the flex classes were moved accordingly.