Updated the post category layouts to be stylistically consistent
An article _include file was introduced to handle formatting of individual posts. The existing category layouts were then updated to use this _include. By putting the distinguishing logic in the individual category layouts and using a common _include, the categories are subtley different but the pages within a specific category will remain self-consistent even when posts are copied in from a different category type, which may introduce post variables that are extraneous in the copied-into category.